Clean up solution, remove duplicate files
