--- modal/src/App.css
@@ -1,15 +1,0 @@
-html,
-body {
-  height: 100%;
-}
-
-body {
-  background: #eee;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.App {
-  text-align: center;
-}
