--- modal/src/components/Modal.css
@@ -1,47 +1,0 @@
-.modal {
-  width: 500px;
-  display: block;
-  background: white;
-  border: 1px solid #ccc;
-  box-shadow: -2rem 2rem 2rem rgba(0, 0, 0, 0.2);
-  position: fixed;
-}
-
-.modal header {
-  display: flex;
-}
-.modal .body {
-  padding: 1rem;
-}
-._modal-close {
-  position: absolute;
-  top: 0;
-  right: 0;
-  padding: 0.5em;
-  line-height: 1;
-  background: #f6f6f7;
-  border: 0;
-  box-shadow: 0;
-  cursor: pointer;
-}
-
-._modal-close-icon {
-  width: 25px;
-  height: 25px;
-  fill: transparent;
-  stroke: black;
-  stroke-linecap: round;
-  stroke-width: 2;
-}
-
-._hide-visual {
-  border: 0 !important;
-  clip: rect(0 0 0 0) !important;
-  height: 1px !important;
-  margin: -1px !important;
-  overflow: hidden !important;
-  padding: 0 !important;
-  position: absolute !important;
-  width: 1px !important;
-  white-space: nowrap !important;
-}
