+++ src/components/modal/Modal.css
@@ -5,19 +5,43 @@
   border: 1px solid #ccc;
   box-shadow: -2rem 2rem 2rem rgba(0, 0, 0, 0.2);
   position: fixed;
-  z-index: 1; /* Sit on top */
+}
+
+.modal header {
+  display: flex;
 }
 .modal .body {
   padding: 1rem;
 }
-.modal .footer {
-  border-top: 1px solid #ccc;
-  background: #eee;
-  padding: 0.5rem 1rem;
+._modal-close {
+  position: absolute;
+  top: 0;
+  right: 0;
+  padding: 0.5em;
+  line-height: 1;
+  background: #f6f6f7;
+  border: 0;
+  box-shadow: 0;
+  cursor: pointer;
 }
-.modal .footer button {
-  background: #76aff0;
-  border-radius: 5px;
-  padding: 0.5rem 1rem;
-  font-size: 0.8rem;
+
+._modal-close-icon {
+  width: 25px;
+  height: 25px;
+  fill: transparent;
+  stroke: black;
+  stroke-linecap: round;
+  stroke-width: 2;
+}
+
+._hide-visual {
+  border: 0 !important;
+  clip: rect(0 0 0 0) !important;
+  height: 1px !important;
+  margin: -1px !important;
+  overflow: hidden !important;
+  padding: 0 !important;
+  position: absolute !important;
+  width: 1px !important;
+  white-space: nowrap !important;
 }
