+++ src/components/modal/Modal.js
@@ -1,11 +1,12 @@
 import React from 'react';
 import ModalContent from './ModalContent';
 import Button from '../button/button';
+import tabbable from '../tabbable';
 class Modal extends React.Component {
   constructor() {
     super();
     this.state = {
-      show: false
+      show: false,
     };
   }
   showModal = () => {
@@ -16,26 +17,51 @@ class Modal extends React.Component {
   closeModal = () => {
     this.setState({ show: false });
     this.TriggerButton.focus();
+    this.props.onClose();
+  };
+
+  onKeyDown = (e) => {
+    let focusableElements = [];
+    focusableElements = tabbable(document.getElementById('modal'));
+    const firstFocusableElement = focusableElements[0];
+    const lastFocusableElement =
+      focusableElements[focusableElements.length - 1];
+    if (e.keyCode === 27) {
+      this.closeModal();
+    }
+    if (e.keyCode === 9 || e.key === 'Tab') {
+      if (e.shiftKey && document.activeElement === firstFocusableElement) {
+        e.preventDefault();
+        lastFocusableElement.focus();
+      } else if (
+        !e.shiftKey &&
+        document.activeElement === lastFocusableElement
+      ) {
+        e.preventDefault();
+        firstFocusableElement.focus();
+      }
+    }
   };
   render() {
     return (
-      <React.Fragment>
+      <>
         <Button
           showModal={this.showModal}
-          buttonRef={n => (this.TriggerButton = n)}
-          triggerText={this.props.modalProps.triggerText}
+          buttonRef={(n) => (this.TriggerButton = n)}
+          triggerText={this.props.modalProps.title}
         />
         {this.state.show ? (
           <ModalContent
-            modalRef={n => (this.modal = n)}
-            buttonRef={n => (this.closeButton = n)}
+            buttonRef={(n) => (this.closeButton = n)}
             closeModal={this.closeModal}
-            content={this.props.modalContent}
-          />
+            onKeyDown={this.onKeyDown}
+          >
+            {this.props.children}
+          </ModalContent>
         ) : (
-          <React.Fragment />
+          <></>
         )}
-      </React.Fragment>
+      </>
     );
   }
 }
